Rename app and CLI
